+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/plants/vegetation/PoisonIvyBlock.java
@@ -7,19 +7,16 @@ import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.VineBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.NotNull;
 
 public class PoisonIvyBlock extends VineBlock {
-    public PoisonIvyBlock(Properties properties) {
-        super(properties);
-    }
 
     public PoisonIvyBlock() {
-        super(BlockBehaviour.Properties.of().sound(SoundType.GRASS).strength(0.2f).randomTicks().noCollission());
+        super(BlockBehaviour.Properties.copy(Blocks.VINE));
     }
 
     @Override