+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.5.3
+- Make Poison Ivy and Skyris Vine copy Vanilla Vine Properties
 
 # 1.5.2
 - Remove Milk from Allium Oddion Soup Recipe, Change Nutrition value from 9 to 8